Move "Contracts on Project Docs" features into properly named modules
The "Service Desk" modules are actually a use case for the feature
adding related Contract information to Project Tasks and Issues.

Moving those features to properly named modules will make it easier
for their adoption and reuse.

For backward compatibility the Service Desk modules are kept, although
with little or no additional features at the moment.