[ADD] project_double_alias
========================
Double alias in projects
========================

Odoo by default allows to define an alias to receive mails, but you can only
select to create tasks or claims related to the project, but not both.

This module makes possible to create an alias for each.

Installation
============

On installation time, this module will move all the existing project aliases
that point to issues to the second alias. If the configured alias is for tasks,
an alias for issues is searched in the system anyway for being linked.

This operation will be reverted back when uninstalling (move second aliases to
the first one when no alias is defined for tasks).

In case you have projects which have configured both aliases for tasks and
issues, both are preserved on uninstallation of this module, but you will
only have accessed to the task one.

Configuration
=============

* Enable developer mode.
* Go to *Settings > Parameters > System parameters*.
* Create or edit the parameter with the name "mail.catchall.domain", and put
  as value the domain of your mail server.

Usage
=====

* Go to a project
* Access the "E-mails" tab.
* Define a mail alias for the tasks and a mail alias for issues