[IMP] project_task_dependency: Improve performance on task form
Depending tasks are very expensive to be computed, so we remove them
from the form view as they are useless.

We need also to split the compute method for each field for not suffering the
same problem with the needed field. Any way, this doesn't add anything, as
each field is computed isolatedly.

Finally, a bit of tweaking has been done for improving a bit computation of
dependencies.