[FIX] project_category can now be migrated from 11.0 to 12.0
In commit 26a7e26481bc35306ed9aacfb0d2f26bbf78a37d the view project_type_search
of project.project was renamed to project_search and there was a new view
introduced with the same XML ID, but for a different model, project.type. This
means that the inherit_id field of that view still contains the old value of
project.view_project_project_filter and obviously that view shouldn't inherit
any other view. Thus the inherit_id is explicitly set to False.