project_milestone: fix incorrect calculation of milestone progress
previously the variables total_tasks_count and closed_tasks_count were defined before looping through the recordset and then incremented during the loop. The result was that if multiple milestones were being calculated at the same time the progress was being calculated on the basis of all tasks within all milestones in the recordset. This commit moves those variables to inside the loop so that they are reset for every milestone, returning accurate progress values for each milestone.

This commit also modifies the unit test to test this scenario, as well as testing the progress on a milestone without any tasks.