Auto fold task on Personal Stage when is closed
This module adds the behavior to change the Personal Stage of the task to the first Folded one when the task reachs an stage with "Mark as Done".
If the Personal Stage is already folded, does nothing.

Simplifies the behavior of closing a task, go back to Personal Stages and move it to the Pesonal Done stage.
Chooses Personal Stages with the same name before Closing stages or Folded stages.