[FIX] project_task_code: in place argument modification
The overload of project.task::create  in the project_task_code would
mutate the dictionaries passed in the vals_list argument. This is a bad
practice which can have unintended side effects in caller code.

We fix this by creating a new dictionary containing the additional field
value and passing this dictionary in the call to super().create()