[FIX] project_timesheet_time_control: Avoid texts on buttons
Since v14, `string` attribute of <button> elements are directly shown in
the views as text instead of being a tooltip, and thus, it takes a lot
of space in screen.

And it was not the intended visualization when the modules was
conceived, just a side effect.

Thus, we restore what we expect, moving `string` to `title` for having
the tooltips again.

TT40169