[IMP] project_task_code: enable name search by Task Number
The search functionality has been improved to include the Task Number.
This enhancement is achieved by inheriting the _name_search method to
include the Task Number in the search domain.

The parent method adds the Name to the search domain (1). As a
consequence of this action, when the args parameter is empty in the
inherited method, an incomplete OR term must be added. Without this
addition, the search domain would default to using an AND operator.

Additionally, tests have been added for this new inheritance.

(1) https://github.com/odoo/odoo/blob/4625cb2f/odoo/models.py#L1928