[FIX] project_timeline: Fix unassigned task color
Unassigned tasks are supposed to show up in white instead of default
color, but the previous `== []` comparison is no longer working in Odoo
16 / with web_timeline 16.0.1.0.3.

This is because web_timeline forces string conversion during evaluation
of these color instructions, and an empty `user_ids` is an empty string
in this case.

This does look somewhat odd but fixes the issue short of rewriting
py-eval instructions in web_timeline, which would be a much larger task.