[FIX] project_timeline: Fix demo task dates
This commit fixes inconsistent dates we get when building a demo DB.

This module adds a `date_start` field; this is the one we want to fill
in demo data.
We do not want to fill `date_assign`, which is an automatic field.

This is how this module used to work in v12 (was based on `date_start`).
* v12➔v13 MIG: Odoo removed the base `date_start` field so `date_start` was
  changed to `date_assign` in this module.
* v13➔v14 MIG: `date_start` was added back into this module to restore
  v12 behavior. However, demo data then got left untouched.