[REF] project_timeline: Use dedicated fields for timeline planning
The fields `date_assign` and `date_end` can't be used, as they are
automatically rewritten on certain flow events (user assignation and
stage changed to finished one), so we need dedicated fields for the
planning. A previous change switches `date_assign` to `date_start`, but
it didn't change demo data, and better to use a consistent naming,
prefixing both fields with `planned_`.

This includes the migration scripts for preserving previous data, and
automations to fill planned data from the previous fields. It also
pre-fills planning information from assignation date/close date as a
best effort pre-planning for existing tasks.

TT50618

Co-Authored-By: Pedro M. Baeza <pedro.baeza@tecnativa.com>