[FIX] sale_project_copy_tasks: find the appropriate `project_id` field always
This reverts https://github.com/OCA/project/pull/1416 and adds instead a different, more accurate fix for the problem (which kept yielding different errors).

The main source of problem is that the same view is adding a `project_id` field and editing another `project_id` fields' attributes. For this reason, installing the module didn't give any problems, but updating it was problematic.

The new field appeared above in the view hierarchy than the old one. The xpath for the old one was less specific than necessary, and then found the new one instead. The new one is within a tree subview, so the model is different (`sale.order.line` instead of `sale.order`). Thus the fields exposed to that field are different, and failures happened a bit randomly.

@moduon MT-3426