[16.0][ADD] project_scoped_tags
New module to modify tags behavior for projects and tasks, inspired by GitLab’s tag scoping.

This allow to limit tag availability by projects

Sould be used with "Optional quick create" module to force projects
managers to manage tags from the dedicated view.

This module does not modify the upstream operation of the tags, it should to be installed and
uninstalled without affecting the data already in place